backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.7
python-dotenv==1.0.1
pydantic==2.5.2
pydantic[email]==2.5.2
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
firebase-admin==6.4.0
requests==2.31.0
beautifulsoup4==4.12.2
pymupdf==1.23.6
google-generativeai==0.4.0
lxml==4.9.3